// src/client/models/impact.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Navitia API - SNCF
 * Complete API documentation for Navitia/SNCF public transportation API. Based on https://doc.navitia.io
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ImpactApplicationPeriodsInner } from './impact-application-periods-inner';
// May contain unused imports in some cases
// @ts-ignore
import type { ImpactImpactedStopsInner } from './impact-impacted-stops-inner';
// May contain unused imports in some cases
// @ts-ignore
import type { ImpactPtObject } from './impact-pt-object';
// May contain unused imports in some cases
// @ts-ignore
import type { Severity } from './severity';

export interface Impact {
    'severity'?: Severity;
    'application_periods'?: Array<ImpactApplicationPeriodsInner>;
    'pt_object'?: ImpactPtObject;
    /**
     * List of impacted stops
     */
    'impacted_stops'?: Array<ImpactImpactedStopsInner>;
    /**
     * Affected object (deprecated, use pt_object)
     */
    'object'?: object;
    /**
     * Affected object type (deprecated, use pt_object.embedded_type)
     */
    'object_type'?: string;
}